refactor: Created file ClassComparisson.tsx for better modularity
